Split LTX2.3 image-to-video pipeline
