+++ tests/diffusion/models/ltx2/test_ltx2_3_pipeline.py
@@ -75,7 +75,8 @@ def _resolve_request_inputs_for_test(pipe, req):
 
 class TestLTX23RequestParsing:
     def test_t2v_and_i2v_share_request_input_resolution(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline, LTX23Pipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
         from vllm_omni.diffusion.request import OmniDiffusionRequest
         from vllm_omni.diffusion.worker.request_batch import DiffusionRequestBatch
         from vllm_omni.inputs.data import OmniDiffusionSamplingParams
@@ -275,13 +276,14 @@ class TestPipelineIndependence:
 
 class TestLTX23ImageToVideoPipeline:
     def test_ltx23_i2v_pipeline_reuses_ltx23_semantics(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline, LTX23Pipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         assert issubclass(LTX23ImageToVideoPipeline, LTX23Pipeline)
         assert LTX23ImageToVideoPipeline.support_image_input is True
 
     def test_ltx23_i2v_rejects_multi_image_prompt_list(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         image = object()
 
@@ -290,7 +292,7 @@ class TestLTX23ImageToVideoPipeline:
             LTX23ImageToVideoPipeline._resolve_single_prompt_image([object(), object()])
 
     def test_ltx23_i2v_additional_image_resolution_is_tensor_safe(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         image = torch.zeros(1, 3, 4, 4)
         additional = {
@@ -302,8 +304,8 @@ class TestLTX23ImageToVideoPipeline:
         assert LTX23ImageToVideoPipeline._resolve_additional_image(additional) is image
 
     def test_ltx23_i2v_packed_latents_are_not_noised(self, monkeypatch):
-        import vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 as ltx23
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline
+        import vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video as ltx23_i2v
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -315,7 +317,7 @@ class TestLTX23ImageToVideoPipeline:
         def fake_randn_tensor(shape, generator=None, device=None, dtype=None):
             raise AssertionError("packed I2V latents should not be noised")
 
-        monkeypatch.setattr(ltx23, "randn_tensor", fake_randn_tensor)
+        monkeypatch.setattr(ltx23_i2v, "randn_tensor", fake_randn_tensor)
 
         latents = torch.tensor([[[10.0, 11.0], [20.0, 21.0], [30.0, 31.0]]])
 
@@ -337,8 +339,8 @@ class TestLTX23ImageToVideoPipeline:
 
     def test_ltx23_i2v_5d_latents_noise_preserves_conditioning_frame(self, monkeypatch):
         import vllm_omni.diffusion.models.ltx2.pipeline_ltx2 as ltx2
-        import vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 as ltx23
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline
+        import vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video as ltx23_i2v
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -355,7 +357,7 @@ class TestLTX23ImageToVideoPipeline:
         def fake_randn_tensor(shape, generator=None, device=None, dtype=None):
             return torch.ones(shape, device=device, dtype=dtype)
 
-        monkeypatch.setattr(ltx23, "randn_tensor", fake_randn_tensor)
+        monkeypatch.setattr(ltx23_i2v, "randn_tensor", fake_randn_tensor)
         monkeypatch.setattr(ltx2, "randn_tensor", fake_randn_tensor)
 
         latents = torch.tensor([[[[[10.0]], [[20.0]], [[30.0]]], [[[11.0]], [[21.0]], [[31.0]]]]])
@@ -377,7 +379,7 @@ class TestLTX23ImageToVideoPipeline:
         torch.testing.assert_close(out, torch.tensor([[[10.0, 11.0], [1.0, 1.0], [1.0, 1.0]]]))
 
     def test_ltx23_i2v_video_step_preserves_conditioning_frame(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -1267,29 +1269,29 @@ class TestPostProcessFunction:
         assert "audio_sample_rate" not in result
 
 
-class TestReExportModule:
-    """Test that pipeline_ltx2_3_image2video.py correctly re-exports."""
+class TestImageToVideoModule:
+    """Test that pipeline_ltx2_3_image2video.py exposes I2V entry points."""
 
     def test_i2v_classes_importable(self):
-        """I2V classes must be importable from the re-export module."""
+        """I2V classes must be importable from the implementation module."""
         from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 
         assert LTX23ImageToVideoPipeline is not None
 
     def test_post_process_func_importable(self):
-        """get_ltx2_post_process_func must be importable from re-export module."""
+        """get_ltx2_post_process_func must be importable for registry lookup."""
         from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import get_ltx2_post_process_func
 
         assert callable(get_ltx2_post_process_func)
 
-    def test_i2v_classes_are_same_as_direct_import(self):
-        """Re-exported classes must be the same objects as direct imports."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23ImageToVideoPipeline as Direct
+    def test_i2v_class_matches_package_export(self):
+        """Package export must point at the I2V implementation module."""
+        from vllm_omni.diffusion.models.ltx2 import LTX23ImageToVideoPipeline as PackageExported
         from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import (
-            LTX23ImageToVideoPipeline as ReExported,
+            LTX23ImageToVideoPipeline as ModuleExported,
         )
 
-        assert Direct is ReExported
+        assert PackageExported is ModuleExported
 
 
 class TestInitExports:
