+++ vllm_omni/diffusion/models/ltx2/__init__.py
@@ -10,10 +10,8 @@ from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import (
     get_ltx2_post_process_func,
     load_transformer_config,
 )
-from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import (
-    LTX23ImageToVideoPipeline,
-    LTX23Pipeline,
-)
+from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
+from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_image2video import (
     LTX2I2VDMD2Pipeline,
     LTX2ImageToVideoPipeline,
